Add isMeasureFormatInPercent to sdk-model
Use this in sdk-ui-ext.
Also some TODOs were extended with JIRA tickets, and a typing was improved.

JIRA: RAIL-2847
